+++ lib/developerTestAccounts.ts
@@ -8,39 +8,23 @@ import {
   DeveloperTestAccount,
   FetchDeveloperTestAccountsResponse,
 } from '../types/developerTestAccounts';
-import { throwError } from '../errors/standardErrors';
 
 export async function createDeveloperTestAccount(
   accountId: number,
   accountName: string
 ): Promise<DeveloperTestAccount> {
-  try {
-    const resp = await _createDeveloperTestAccount(accountId, accountName);
-    return resp;
-  } catch (err) {
-    throwError(err);
-  }
+  return _createDeveloperTestAccount(accountId, accountName);
 }
 
 export async function deleteDeveloperTestAccount(
   accountId: number,
   testAccountId: number
 ): Promise<void> {
-  try {
-    const resp = await _deleteDeveloperTestAccount(accountId, testAccountId);
-    return resp;
-  } catch (err) {
-    throwError(err);
-  }
+  return _deleteDeveloperTestAccount(accountId, testAccountId);
 }
 
 export async function fetchDeveloperTestAccounts(
   accountId: number
 ): Promise<FetchDeveloperTestAccountsResponse> {
-  try {
-    const resp = await _fetchDeveloperTestAccounts(accountId);
-    return resp;
-  } catch (err) {
-    throwError(err);
-  }
+  return _fetchDeveloperTestAccounts(accountId);
 }
