+++ lib/fs.ts
@@ -3,7 +3,6 @@ import path from 'path';
 
 import { STAT_TYPES } from '../constants/files';
 import { StatType, FileData } from '../types/Files';
-import { throwError } from '../errors/standardErrors';
 
 export function getFileInfoAsync(dir: string, file: string): Promise<FileData> {
   return new Promise((resolve, reject) => {
@@ -69,6 +68,6 @@ export async function walk(dir: string): Promise<Array<string>> {
     .then(processFiles)
     .then(flattenAndRemoveSymlinks)
     .catch(err => {
-      throwError(err);
+      throw err;
     });
 }
