+++ lib/oauth.ts
@@ -1,7 +1,6 @@
 import OAuth2Manager from '../models/OAuth2Manager';
 import { AUTH_METHODS } from '../constants/auth';
 import { FlatAccountFields } from '../types/Accounts';
-import { throwError } from '../errors/standardErrors';
 import { logger } from './logger';
 import { getAccountIdentifier } from '../utils/getAccountIdentifier';
 import { updateAccountConfig, writeConfig } from '../config';
@@ -39,14 +38,10 @@ export function getOauthManager(
 
 export function addOauthToAccountConfig(oauth: OAuth2Manager) {
   logger.log(i18n(`${i18nKey}.addOauthToAccountConfig.init`));
-  try {
-    updateAccountConfig({
-      ...oauth.account,
-      authType: AUTH_METHODS.oauth.value,
-    });
-    writeConfig();
-    logger.success(i18n(`${i18nKey}.addOauthToAccountConfig.success`));
-  } catch (err) {
-    throwError(err);
-  }
+  updateAccountConfig({
+    ...oauth.account,
+    authType: AUTH_METHODS.oauth.value,
+  });
+  writeConfig();
+  logger.success(i18n(`${i18nKey}.addOauthToAccountConfig.success`));
 }
